Add files via upload
added new operators and resolved warnings

+++ vpr/src/draw/draw.h
@@ -79,7 +79,7 @@ void toggle_placement_macros(GtkWidget* /*widget*/, gint /*response_id*/, gpoint
 
 ezgl::color get_block_type_color(t_physical_tile_type_ptr type);
 
-void breakpoint_info_window(std::string bpDescription, int move, int temp, int block, int net);
+void breakpoint_info_window(std::string bpDescription, int move, float temp, int block, int net);
 
 #endif /* NO_GRAPHICS */
